+++ src/Test/Models/TestSection.php
@@ -25,7 +25,7 @@ class TestSection extends Section
     public const int TYPE_LINK_LIST = 5;
 
     #[\Override]
-    public static function getTypes(): array
+    public function getTypes(): array
     {
         return [
             SectionType::make(self::TYPE_HEADLINE)